fix: Address self review comments
