+++ nisystemlink/clients/auth/models/_auth_models.py
@@ -1,5 +1,4 @@
-# generated by datamodel-codegen:
-#   filename:  <stdin>
+"""Models utilized for Auth in SystemLink."""
 
 from __future__ import annotations
 
@@ -369,3 +368,18 @@ class PolicyTemplate(JsonModel):
     """
     A list of statements defining the actions the user can perform on a resource
     """
+
+
+class AuthInfo(JsonModel):
+    """Information about the authenticated caller."""
+
+    user: Optional[User]
+    """Details of authenticated caller"""
+    org: Optional[Org]
+    """Organization of authenticated caller"""
+    workspaces: Optional[List[Workspace]]
+    """List of workspaces the authenticated caller has access"""
+    policies: Optional[List[AuthPolicy]]
+    """List of policies for the authenticated caller"""
+    properties: Optional[Dict[str, str]] = Field(None, example={"key1": "value1"})
+    """A map of key value properties"""
